requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
python-dotenv==1.0.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
isort==5.12.0
python-decouple==3.8 